requirements.txt
django==1.11.4
psycopg2==2.7.3
pytz==2017.2
dj-database-url==0.4.2
python-magic==0.4.13
